backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
icalendar==5.0.13
apscheduler==3.10.4
aiosmtplib==3.0.1
python-dotenv==1.0.1
pydantic-settings==2.4.0
